fe:  add new feature
- [x] 1. Gnebe로그수정 / 품번 띄어쓰기로
- [x] 2. Gnebe file 저장해서 한번 올라온 상품 다시 안올라게끔 loading시 데이터 올려두기
- [x] 3. Gnebe  프로그램 실행 on/off and

+++ src/main/java/com/example/monitor/infra/s3/S3UploaderService.java
@@ -3,12 +3,15 @@ package com.example.monitor.infra.s3;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
+import java.net.URL;
 
 @Service
 @RequiredArgsConstructor
@@ -17,24 +20,35 @@ public class S3UploaderService {
 
     private final AmazonS3Client amazonS3Client;
 
+    @Setter
+    @Getter
+    private boolean isAllowedUpload = true;
+
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
 
-    public String uploadImage(File file, String fileName) {
+    public String uploadImageOrNull(File file, String fileName) {
 
         assert (file != null);
         assert (fileName != null);
-        amazonS3Client.putObject(
-                new PutObjectRequest(bucket, fileName, file)
-                        .withCannedAcl(CannedAccessControlList.PublicRead)    // PublicRead 권한으로 업로드 됨
-        );
+
+        if (!isAllowedUpload) {
+            file.delete();
+            return null;
+        }
+        //상품이 있으면 재업로드 안해도됨
         String accessUrl = amazonS3Client.getUrl(bucket, fileName).toString();
 
-        boolean delete = file.delete();
-
-        if (delete) {
-            log.info("upload 후 로컬 파일삭제");
+        if (accessUrl == null) {
+            amazonS3Client.putObject(
+                    new PutObjectRequest(bucket, fileName, file)
+                            .withCannedAcl(CannedAccessControlList.PublicRead)    // PublicRead 권한으로 업로드 됨
+            );
+            accessUrl = amazonS3Client.getUrl(bucket, fileName).toString();
         }
+
+        file.delete();
+
 
         return accessUrl;
     }
